backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
opencv-python
numpy
pillow
gunicorn
whitenoise
